epi_monitor/requirements.txt
ultralytics
opencv-python
python-dotenv
requests
pydantic
